Reorganize commands under 'ar' for Argentina
Drop "saij" for "ar", since we'll just use that for all of Argentina's norms.

This involves also reorganizing commands and services configuration, offering guidance on how to do it in the future for other countries.

Add test on how to project yaml front-matter from a document, as well as full metadata in yaml format.

We also add empty/hidden anchors for each article when converting to markdown for easier deep-linking.