Add support for enumerating all doc types and jurisdictions
A new [DisplayValue] is introduced to allow converting from enum to display value (including spaces and accents as needed) for URL building purposes (SAIJ uses spaces and accents in paths :/ ).

Add tests that exercise the enumeration across all doc types and all jurisdictions for law and decree. This also drove a fix to an issue where we were appending jurisdiction when it was invalid (i.e. other doc types that don't support that filter), which resulted in empty enumerations.