Improve support for enumeration of all supported content types
The types of norms supported in TipoNorma fit under the Document Type = Legislation on the SAIJ website. These are the types of docs that have the metadata we support in our abstract doc enumeration.

The other types would be useful to support but would require updated data models too, and likely different API overloads, so we just disable them for now: when enumerating, if we find a content-type that can't be parsed into our enum, we simply skip it.