Expose SAIJ ID in addition to document UUID
When searching, the results contain just latter, whereas full document loading contains both. The former is better for shorter links and filenames, while the latter is required to retrieve full document data.

We improve the FetchAsync method to support fetching by the shorter SAIJ ID too by performing a quick search. Using these IDs would then necessarily be a bit slower, but the benefit is a better UEX in some scenarios.

We update the markdown rendering to include selected properties as-is so that the resulting dictionary can be merged as-is with the front-matter at the end of the document without conflicts or duplication.

We also add both IDs at the end the front-matter since in GitHub it will go at the end of a scrollbar anyway and not be too much of a rendering issue, while we can use them to avoid reading entire file contents just to retrieve those values.

The included timestamp will also help in the sync process to come.