Add basic cross-repository sync interfaces and core implementation
This initial cut focuses on getting the API right. We rename back the Uuid > Id since that's more agnostic to specific repositories, and introduce Alias for the SAIJ ID instead. The alias is still used for the web urls, and the ID for data urls.

The sync algorithm enumerates the source repo, gets the timestamp from the target repo, and acts accordingly by creating or updating the content in the target repo if it doesn't exist or doesn't match the source repo timestamp.

For the file repo, we check the timestamp as an optimization nevertheless, in case the caller of the repo API didn't previosly check the timestamp. Originally,we had only this method and no separte GetTimestamp, but since fetching the document content from SAIJ requires another HTTP request, we wanted to avoid it if the target file/timestamp already matched, for performance reasons.