Big refactoring to simplify and clarify usage and supoorted data
We were trying to show too much data on the search result object, but since seaches return disparate documents depending on the content source, it was getting complicated to unify.

In fact, some docs don't even have timestamps, so we need to deal with that too.

Normalizing JSON data was also cumbersome and now that search result and loaded document are two distinct objects, we introduce a reusable component `NormalizedWebDocument` to contain the logic we had before in the single Document class.

The IContentInfo is now barely used, and the old SaijClient-based repository is also unnecessary since we need to do specific work before comparing timestamps (load the full doc).