Fix issue with negative timestamps
Turns out that JQ on Linux does support negative timestamps, but not Windows.
This is important because there are many docs with dates predating the unix
epoc.

So we switch to number of seconds since 1/1/1 for simplicity, and handle that
uniformly across search and document metadata.

We also add error doc persistence so we can later use a regression test to
re-load them if they occur.

The way this would work, is the target folder will end up with a .openlaw/errors
folder with the problematic docs. We can pull docs from there for reproducing,
fix the bug, and then commit the deletion of the error docs once we get a fix
deployed.