Add support for reading the normative references
For now, we just collect the references as belonging to the doc as a whole. When serializing as YAML front-matter, we only persist the count of linked norms, to keep the markdown small. The full data about normative references would still require full data fetch in JSON format and further analysis anyway, since the references apply to specific sections of a norm.

We took the chance to remove SharpYaml and avoid confusion about which lib we're using (for now?) for our YAML needs.