Refactor sync command and file format
Rather than trying to squeeze metadata into the main markdown file, keep it mostly clean so we can upload as-is to a vector store for semantic query/search. Keep the full source data document separate instead under a subfolder.

This also allows us to keep the shorter, SAIJ-specific document ID/alias, since repo operations index on the data, rather than the markdowns.

This simplifies things given the non-goal of making these docs human editable (for now?), and therefore we don't have a need to keep its metadata in sync with the actual document data (since it's all updated by tools).

So we remove the end(front)-matter in markdown now.