Expose ContentType as a sync filter to avoid empty "novedades"
Content type "novedad" is typically empty and a pointer to actual legislation.

This change allow sync command to optionally receive the content type to sync.
We'll default to just legislation now by default, but we allow explicitly fetching news.

Note that there will be a mismatch between the initial query totals and the actual total
being sync'ed afterwards: this will likely be due to the skipping of `novedad` from the
sync process. It might be a bit confusing since we're not calling that action in the
progress, but it seems low-importance anyway.