Refactor for reuse sync behavior from sync item, skip timestamps-only
As part of improving the detection of timestamps-only changes, we also refactor key parts of SyncCommand (particularly SyncAction) so it can be reused from SyncItemCommand. This involves using as item ID the infojus ID rather than the uuid, so we can share the rest of the code based on search results.

We rename OpenLaw > OpenLaw.Commands (in preparation to moving more non-CLI code to the Core project).

Added non-CLI code to the new OpenLaw.Core (perhaps we'll rename that to just OpenLaw down the road).